requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
celery[redis]>=5.3.0
redis>=5.0.0
python-jobspy>=1.1.0
pdfplumber>=0.10.0
playwright>=1.40.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
